Improved script

+++ bin/AnalysisJetsSmrBTagRegFSR.sh
@@ -7,15 +7,20 @@ CORRECTION_LEVELS=(nothing only_smearing smearing_btag smearing_btag_regression 
 TRIGGER_FLAGS=("true" "false")
 PFILE=Parameters.cc
 
-set -v
 cmsenv
 hash -r
+mkdir -p $OUTPUT_DIR
 
 for corr in ${CORRECTION_LEVELS[@]} ; do
     for trig in ${TRIGGER_FLAGS[@]} ; do
-        echo "const CorrectionLevel correction_level = CorrectionLevel::${corr};" > $PFILE
-        echo "const bool apply_trigger_cut = ${trig};" >> $PFILE
-        scram b
-        AnalysisJetsSmrBTagRegFSR ${OUTPUT_DIR}/${CORRECTION_LEVELS}_${TRIGGER_FLAGS}.root
+        OUT_FILE="${OUTPUT_DIR}/${corr}_${trig}.root"
+        if ! [[ -f ${OUT_FILE} ]] ; then
+            echo "const CorrectionLevel correction_level = CorrectionLevel::${corr};" > $PFILE
+            echo "const bool apply_trigger_cut = ${trig};" >> $PFILE
+            scram b
+            AnalysisJetsSmrBTagRegFSR ${OUT_FILE}
+        else
+            echo -e "\033[0;31mFile already exists. Not overwriting\033[0m"
+        fi
     done
 done
